% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{PyR-package}
\alias{PyR-package}
\title{PyR: Set of Functions to Allow Use of Python in R}
\description{
A bridge between R and Python. Thes package is inspired by rPython
package but adds some additional functionality like dealing with Python
exceptions and converting native Python data types to native R data types.
The pacakge was also inspired by http://gallery.rcpp.org/articles/rcpp-python/ .
}

