% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/python.assign.R
\name{python.assign}
\alias{python.assign}
\title{Assign value to Python variable.}
\usage{
python.assign(var_name, val)
}
\arguments{
\item{var_name}{Variable name string.}

\item{val}{Value to be assigned.}
}
\description{
Assign value to Python variable.
}
\examples{
python.assign("a", 10)
}

