% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/python.call.R
\name{python.call}
\alias{python.call}
\title{Call python function with parameters.}
\usage{
python.call(func, ...)
}
\arguments{
\item{func}{Function name string.}

\item{...}{Optional function parameters, could be names as well.}
}
\value{
Value returned by called function.
}
\description{
Call python function with parameters.
}
\examples{
python.call("str", 100.5)
}

