% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/python.eval.R
\name{python.eval}
\alias{python.eval}
\title{Evaluate Python expression.}
\usage{
python.eval(code)
}
\arguments{
\item{code}{Python code string, evaluated in it's own namespace.}
}
\value{
Result of evaluation.
}
\description{
Evaluate Python expression.
}
\examples{
python.eval("1+1")
}

