% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/python.exec.R
\name{python.exec}
\alias{python.exec}
\title{Execute Python code.}
\usage{
python.exec(code)
}
\arguments{
\item{code}{Python code string.}
}
\description{
Execute Python code.
}
\examples{
python.exec("a = 1 + 2")
}

