% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{python.get}
\alias{python.get}
\title{Get value of Python variable.}
\usage{
python.get(var_name)
}
\arguments{
\item{var_name}{Variable name string.}
}
\value{
Value of the variable.
}
\description{
Get value of Python variable.
}
\examples{
python.get('a')
}

