% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/python.get_class.R
\name{python.get_class}
\alias{python.get_class}
\title{Get Python class.}
\usage{
python.get_class(cls_name, env = parent.frame())
}
\arguments{
\item{cls_name}{Name of the Python class.}

\item{env}{Optional environment, where corresponding R class will be created.}
}
\value{
R class corresponding to Python class.
}
\description{
Get Python class.
}
\examples{
PyStr <- python.get_class("str")
s <- PyStr("hello, {name}")
s$format(name = "world")
}

