% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/python.get_function.R
\name{python.get_function}
\alias{python.get_function}
\title{Get Python function.}
\usage{
python.get_function(fn_name)
}
\arguments{
\item{fn_name}{Function name string.}
}
\value{
Wrapper around Python function.
}
\description{
Get Python function.
}
\examples{
python.exec("def test_fn(a = 3, b = 4):\\n    return a * b")
test_fn <- python.get_function('test_fn')
test_fn(10, 20)
test_fn(a = 10)
test_fn(b = 10)
}

