% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/python.get_object.R
\name{python.get_object}
\alias{python.get_object}
\title{Get Python object.}
\usage{
python.get_object(obj_name, env = parent.frame())
}
\arguments{
\item{obj_name}{Name of the Python object.}

\item{env}{Optional environment, where corresponding R object will be created.}
}
\value{
R object corresponding to Python object.
}
\description{
Get Python object.
}
\examples{
PyStr <- python.get_object("str")
s <- PyStr("hello, {name}")
s$format(name = "world")
}

