% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/python.call.R
\name{python.call}
\alias{python.call}
\title{Call python function with parameters.}
\usage{
python.call(func, ..., .saveTo = "pyr._", .getResults = TRUE)
}
\arguments{
\item{func}{Function name string.}

\item{...}{Optional function parameters, could be names as well.}

\item{.saveTo}{Name of the Python variable to save results of the call.}

\item{.getResults}{Wether to return results of the execution.}
}
\value{
Value returned by called function.
}
\description{
Call python function with parameters.
}
\examples{
python.call("str", 100.5)
}

